Added way to exit menus such as "Make file, folder and etc" with <0> symbol

+++ src/menu.cpp
@@ -28,56 +28,69 @@ void menu()
         {
             case 1: // Создание файла
             {
-                std::string question = "Введите название создаваемого файла";
+                std::string question = "Введите название создаваемого файла, для выхода введите <0>";
                 std::string file = askName(question);
-                File* BuffFile = new File;
-                BuffFile->name = file;
-                MainSys.touch(BuffFile); 
+                if (file != "0")
+                {
+                    File* BuffFile = new File;
+                    BuffFile->name = file;
+                    MainSys.touch(BuffFile);                     
+                }
 
                 break;
             }
             case 2: // Создание папки
             {
-                std::string question = "Введите название создаваемой папки";
+                std::string question = "Введите название создаваемой папки, для выхода введите <0>";
                 std::string folder = askName(question);
-                Folder* BuffFolder = new Folder;
-                BuffFolder->name = folder;
-                MainSys.mkdir(BuffFolder);
+                if (folder != "0")
+                {
+                    Folder* BuffFolder = new Folder;
+                    BuffFolder->name = folder;
+                    MainSys.mkdir(BuffFolder);                    
+                }
 
                 break;
             }
             case 3: // Удаление файла
             {
-                std::string question = "Введите название удаляемого файла";
+                std::string question = "Введите название удаляемого файла, для выхода введите <0>";
                 std::string file = askName(question);
-                File* BuffFile = static_cast<File*>(MainSys.findElem(file));
+                if (file != "0")
+                {
+                    File* BuffFile = static_cast<File*>(MainSys.findElem(file));
 
-                if (BuffFile != nullptr)
-                    MainSys.rm(BuffFile);
-                else
-                    std::cout << "В текущей директории нет файла с данным названием" << std::endl << std::endl;
+                    if (BuffFile != nullptr)
+                        MainSys.rm(BuffFile);
+                    else
+                        std::cout << "В текущей директории нет файла с данным названием" << std::endl << std::endl;                   
+                }
 
                 break;        
             }
             case 4: // Удаление папки
             {
-                std::string question = "Введите название удаляемой папки файла";
+                std::string question = "Введите название удаляемой папки файла, для выхода введите <0>";
                 std::string folder = askName(question);
-                Folder* BuffFolder = static_cast<Folder*>(MainSys.findElem(folder));
+                if (folder != "0")
+                {
+                    Folder* BuffFolder = static_cast<Folder*>(MainSys.findElem(folder));
 
-                if (BuffFolder != nullptr)
-                    MainSys.rmdir(BuffFolder);
-                else
-                    std::cout << "В текущей директории нет папки с данным названием" << std::endl << std::endl;
+                    if (BuffFolder != nullptr)
+                        MainSys.rmdir(BuffFolder);
+                    else
+                        std::cout << "В текущей директории нет папки с данным названием" << std::endl << std::endl;                    
+                }
 
                 break;
             }
             case 5: // Изменение директории
             {
                 std::cout << MainSys.fullPath << std::endl;
-                std::string question = "Введите название папки, в которую хотите войти,\nдля выхода из текущей папки введите <..>";
+                std::string question = "Введите название папки, в которую хотите войти,\nдля выхода из текущей папки введите <..>,\nдля выхода введите <0>";
                 std::string folder = askName(question);
-                MainSys.cd(folder);
+                if (folder != "0")
+                    MainSys.cd(folder);
 
                 break;
             }
